src/requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
base58==1.0.0
bitcoin==1.1.39
bitmerchant==0.1.8
blockcypher==1.0.73
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
dj-database-url==0.5.0
Django==2.0
django-authtools==1.6.0
django-extensions==2.0.7
django-filter==2.0.0.dev1
django-phonenumber-field==2.0.0
django-widget-tweaks==1.4.2
ecdsa==0.13
gunicorn==19.8.1
idna==2.7
phonenumberslite==8.9.9
Pillow==5.1.0
psycopg2==2.7.4
pyasn1==0.4.3
pycparser==2.18
pycrypto==2.6.1
python-dateutil==2.2
python-decouple==3.1
pytz==2018.4
requests==2.4.3
six==1.11.0
urllib3==1.23
